Begin gemaakt met omzetten naar MVVM
