+++ RondjeBreda.ViewModels/HomePageViewModel.cs
@@ -1,8 +1,11 @@
 ﻿using System.Collections.ObjectModel;
 using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+using Microsoft.Maui.Controls.Maps;
 using RondjeBreda.Domain.Interfaces;
 using RondjeBreda.Domain.Models.DatabaseModels;
 using Location = RondjeBreda.Domain.Models.DatabaseModels.Location;
+using Microsoft.Maui.Maps;
 
 namespace RondjeBreda.ViewModels;
 
@@ -16,14 +19,23 @@ public partial class HomePageViewModel : ObservableObject
     public IGeolocation geolocation;
     public IMapsAPI mapsAPI;
     private string route;
-    // private Map map; TODO
     private bool routePaused;
     public double userLat, userLon;
+    private List<Domain.Models.DatabaseModels.Location> routePoints;
+    private Location nextLocation;
+    private int indexRoute = 0;
 
     [ObservableProperty] private ObservableCollection<Route> routes;
+    [ObservableProperty] private ObservableCollection<Pin> pins;
+    [ObservableProperty] private ObservableCollection<MapElement> mapElements;
+    [ObservableProperty] public MapSpan mapSpan;
+
 
     public HomePageViewModel(IDatabase database, IPreferences preferences, IMapsAPI mapsAPI, IGeolocation geolocation)
     {
+        pins = new ObservableCollection<Pin>();
+        mapElements = new ObservableCollection<MapElement>();
+
         this.database = database;
         this.preferences = preferences;
         this.mapsAPI = mapsAPI;
@@ -42,8 +54,149 @@ public partial class HomePageViewModel : ObservableObject
     {
         this.userLat = e.Location.Latitude;
         this.userLon = e.Location.Longitude;
+
+        if (nextLocation == null)
+        {
+            return;
+        }
+        if (Microsoft.Maui.Devices.Sensors.Location.CalculateDistance(userLat, userLon,
+                nextLocation.latitude, nextLocation.longitude, DistanceUnits.Kilometers) <= 0.02)
+        {
+            LocationReached();
+        }
     }
 
+    private void LocationReached()
+    {
+        if (routePoints.Count == 0)
+        {
+            return;
+        }
+
+        this.indexRoute++;
+        if (indexRoute >= routePoints.Count)
+        {
+            indexRoute = 0;
+        }
+        this.nextLocation = routePoints[this.indexRoute];
+        MapElements.Clear();
+        // TODO: picker moet route inladen
+        LoadRoute(new Route());
+        DrawCircleNextLocation();
+        SetMapSpan();
+    }
+
+    private async void LoadRoute(Route selectedRoute)
+    {
+        // Punten van de geselecteerde route laden
+        this.routePoints = await LoadPoints();
+        if (nextLocation == null)
+        {
+            this.nextLocation = routePoints[0];
+        }
+
+        // Punten toevoegen aan de map
+        Pins.Clear();
+        foreach (var location in routePoints)
+        {
+            Pins.Add(new Pin
+            {
+                Label = location.name,
+                Location = new Microsoft.Maui.Devices.Sensors.Location(location.latitude, location.longitude)
+            });
+        }
+
+        // Route genereren tussen de punten
+        foreach (var location in routePoints)
+        {
+            if (routePoints.IndexOf(location) == routePoints.Count - 1)
+            {
+                continue;
+            }
+
+            var locationCount = routePoints.IndexOf(location);
+            var route = await mapsAPI.CreateRoute($"{location.latitude}", $"{location.longitude}",
+                routePoints[locationCount + 1].latitude.ToString(),
+                routePoints[locationCount + 1].longitude.ToString());
+
+            Polyline polyline = new Polyline
+            {
+                StrokeColor = Colors.Blue,
+                StrokeWidth = 12,
+            };
+
+            var locations = mapsAPI.Decode(route.routes[0].overview_polyline.points);
+            foreach (var tempLocation in locations)
+            {
+                polyline.Geopath.Add(new Microsoft.Maui.Devices.Sensors.Location(tempLocation.latitude, tempLocation.longitude));
+            }
+
+            // Lijn toevoegen aan de map
+            MapElements.Add(polyline);
+        }
+
+        // Route to the next point of the route
+        var routeToFirstPoint = await mapsAPI.CreateRoute($"{userLat}", $"{userLon}",
+            $"{nextLocation.latitude}", $"{nextLocation.longitude}");
+        Polyline firstpolyline = new Polyline
+        {
+            StrokeColor = Colors.Chartreuse,
+            StrokeWidth = 12,
+        };
+        var firstlocations = mapsAPI.Decode(routeToFirstPoint.routes[0].overview_polyline.points);
+        foreach (var tempLocation in firstlocations)
+        {
+            firstpolyline.Geopath.Add(new Microsoft.Maui.Devices.Sensors.Location(tempLocation.latitude, tempLocation.longitude));
+        }
+        MapElements.Add(firstpolyline);
+    }
+
+    private void DrawCircleNextLocation()
+    {
+        if (this.routePoints == null || this.routePoints.Count == 0)
+        {
+            return;
+        }
+
+
+        foreach (var location in routePoints)
+        {
+            // TODO: Check if location is visited, first one that don't need to be set
+        }
+
+        Circle circle = new Circle
+        {
+            Center = new Microsoft.Maui.Devices.Sensors.Location(nextLocation.latitude, nextLocation.longitude),
+            Radius = new Distance(20),
+            StrokeColor = Color.FromArgb("#CFffc61e"),
+            StrokeWidth = 8,
+            FillColor = Color.FromArgb("#CFffc61e")
+        };
+
+        MapElements.Add(circle);
+    }
+
+    private void SetMapSpan()
+    {
+        var centerLat = (this.nextLocation.latitude + userLat) / 2;
+        var centerLon = (this.nextLocation.longitude + userLon) / 2;
+
+        var center = new Microsoft.Maui.Devices.Sensors.Location(centerLat, centerLon);
+
+        var distance = Microsoft.Maui.Devices.Sensors.Location.CalculateDistance(new Microsoft.Maui.Devices.Sensors.Location(this.nextLocation.latitude, 
+            this.nextLocation.longitude), center, DistanceUnits.Kilometers);
+
+        MapSpan mapSpan = MapSpan.FromCenterAndRadius(center, Distance.FromKilometers(distance * 1.5));
+    }
+
+    [RelayCommand]
+    private void ImageButtonPressed()
+    {
+        MapElements.Clear();
+        LoadRoute(new Route());
+        DrawCircleNextLocation();
+        SetMapSpan();
+    }
 
     public async Task<List<Location>> LoadPoints()
     {   // TODO: Route component tabel goed ophalen
