+++ RondjeBreda/Pages/MainPage.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.Diagnostics;
+﻿using System.ComponentModel;
+using System.Diagnostics;
 using CommunityToolkit.Mvvm.Input;
 using Microsoft.Maui.Controls.Maps;
 using Microsoft.Maui.Maps;
@@ -18,167 +19,32 @@ namespace RondjeBreda.Pages
     public partial class MainPage : ContentPage
     {
         private HomePageViewModel homePageViewModel;
-        private List<Domain.Models.DatabaseModels.Location> routePoints;
-        private Domain.Models.DatabaseModels.Location nextLocation;
-        private int indexRoute = 0;
-
+        
         public MainPage(HomePageViewModel homePageViewModel)
         {
             InitializeComponent();
             this.homePageViewModel = homePageViewModel;
-            //BindingContext = homePageViewModel;
-            homePageViewModel.geolocation.LocationChanged += (sender, e) =>
-            {
-                if (nextLocation == null)
-                {
-                    return;
-                }
-                if (Location.CalculateDistance(homePageViewModel.userLat, homePageViewModel.userLon,
-                        nextLocation.latitude, nextLocation.longitude, DistanceUnits.Kilometers) <= 0.02)
-                {
-                    LocationReached();
-                }
-            };
+            homePageViewModel.PropertyChanged += MoveToRegion();
+
+            BindingContext = homePageViewModel;
         }
 
-        private void LocationReached()
+        private PropertyChangedEventHandler? MoveToRegion()
         {
-            if (routePoints.Count == 0)
+            if (Map == null || homePageViewModel == null || homePageViewModel.mapSpan == null)
             {
-                return;
+                return null;
             }
-
-            this.indexRoute++;
-            if (indexRoute >= routePoints.Count)
-            {
-                indexRoute = 0;
-            }
-            Debug.WriteLine($"Index: {indexRoute}");
-            this.nextLocation = routePoints[this.indexRoute];
-            Map.MapElements.Clear();
-            LoadRoute(new Route());
-            DrawCircleNextLocation();
-            SetMapSpan();
-        }
-
-        private async void LoadRoute(Route selectedRoute)
-        {
-            // Punten van de geselecteerde route laden
-            this.routePoints = await homePageViewModel.LoadPoints();
-            if (nextLocation == null)
-            {
-                this.nextLocation = routePoints[0];
-            }
-
-            // Punten toevoegen aan de map
-            Map.Pins.Clear();
-            foreach (var location in routePoints)
-            {
-                Map.Pins.Add(new Pin
-                {
-                    Label = location.name,
-                    Location = new Location(location.latitude, location.longitude)
-                });
-            }
-
-            Debug.WriteLine($"Punten: {routePoints.Count}");
-
-            // Route genereren tussen de punten
-            foreach (var location in routePoints)
-            {
-                if (routePoints.IndexOf(location) == routePoints.Count - 1)
-                {
-                    continue;
-                }
-
-                var locationCount = routePoints.IndexOf(location);
-                Debug.WriteLine($"Lat: {location.latitude}, Long: {location.longitude}");
-                var route = await homePageViewModel.mapsAPI.CreateRoute($"{location.latitude}", $"{location.longitude}",
-                    routePoints[locationCount + 1].latitude.ToString(),
-                    routePoints[locationCount + 1].longitude.ToString());
-
-
-                Debug.WriteLine($"Aantal routes: {route.routes.Length}");
-
-                Polyline polyline = new Polyline
-                {
-                    StrokeColor = Colors.Blue,
-                    StrokeWidth = 12,
-                };
-
-                var locations = homePageViewModel.mapsAPI.Decode(route.routes[0].overview_polyline.points);
-                foreach (var tempLocation in locations)
-                {
-                    polyline.Geopath.Add(new Location(tempLocation.latitude, tempLocation.longitude));
-                }
-
-                // Lijn toevoegen aan de map
-                Map.MapElements.Add(polyline);
-            }
-
-            // Route to the next point of the route
-            var routeToFirstPoint = await homePageViewModel.mapsAPI.CreateRoute($"{homePageViewModel.userLat}", $"{homePageViewModel.userLon}",
-                $"{nextLocation.latitude}", $"{nextLocation.longitude}");
-            Polyline firstpolyline = new Polyline
-            {
-                StrokeColor = Colors.Chartreuse,
-                StrokeWidth = 12,
-            };
-            var firstlocations = homePageViewModel.mapsAPI.Decode(routeToFirstPoint.routes[0].overview_polyline.points);
-            foreach (var tempLocation in firstlocations)
-            {
-                firstpolyline.Geopath.Add(new Location(tempLocation.latitude, tempLocation.longitude));
-            }
-            Map.MapElements.Add(firstpolyline);
+            Map.MoveToRegion(homePageViewModel.mapSpan);
+            return null;
         }
 
         private void ImageButton_OnPressed(object? sender, EventArgs e)
         {
+
             Debug.WriteLine("PauseButton!!!");
 
-            Map.MapElements.Clear();
-            LoadRoute(new Route());
-            DrawCircleNextLocation();
-            SetMapSpan();
-        }
-
-        private void DrawCircleNextLocation()
-        {
-            if (this.routePoints == null || this.routePoints.Count == 0)
-            {
-                return;
-            }
-
-
-            foreach (var location in routePoints)
-            {
-                // TODO: Check if location is visited, first one that don't need to be set
-            }
-
-            Circle circle = new Circle
-            {
-                Center = new Location(nextLocation.latitude, nextLocation.longitude),
-                Radius = new Distance(20),
-                StrokeColor = Color.FromArgb("#CFffc61e"),
-                StrokeWidth = 8,
-                FillColor = Color.FromArgb("#CFffc61e")
-            };
-
-            Map.MapElements.Add(circle);
-        }
-
-        private void SetMapSpan()
-        {
-            var centerLat = (this.nextLocation.latitude + homePageViewModel.userLat) / 2;
-            var centerLon = (this.nextLocation.longitude + homePageViewModel.userLon) / 2;
-
-            var center = new Location(centerLat, centerLon);
-
-            var distance = Location.CalculateDistance(new Location(this.nextLocation.latitude, this.nextLocation.longitude), center, DistanceUnits.Kilometers);
-
-            MapSpan mapSpan = MapSpan.FromCenterAndRadius(center, Distance.FromKilometers(distance * 1.5));
-
-            Map.MoveToRegion(mapSpan);
+            
         }
     }
 }
